+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
-## 0.4.0 (unreleased)
+## 0.4.1 (Unreleased)
+
+## 0.4.0 (2023-02-22)
 
 - BREAKING: Rename `node.move()` -> `node.move_to()`
-- New `node.copy_to()`
+- New `tree.copy_to()` and `node.copy_to()`
+- New `tree.children` property
+- Configurable default connector style
 
 
 ## 0.3.0 (2022-08-01)